Refactoring and render error fixing.
