+++ package.json
@@ -9,19 +9,19 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@prisma/client": "^5.19.0",
-    "next": "14.2.5",
+    "@prisma/client": "^5.19.1",
+    "next": "14.2.9",
     "react": "^18",
     "react-dom": "^18",
     "sqlite3": "^5.1.7"
   },
   "devDependencies": {
-    "@types/node": "^20",
+    "@types/node": "^22",
     "@types/react": "^18",
     "@types/react-dom": "^18",
-    "eslint": "^8",
-    "eslint-config-next": "14.2.5",
-    "prisma": "^5.19.0",
+    "eslint": "^9",
+    "eslint-config-next": "14.2.9",
+    "prisma": "^5.19.1",
     "typescript": "^5"
   }
 }
