--- src/error/context.tsx
@@ -1,11 +1,0 @@
-'use client';
-
-import { createContext } from 'react';
-import { ErrorRecord } from './types';
-
-const ErrorRecordContext = createContext<ErrorRecord>({
-  records: [],
-  setRecords: () => {},
-});
-
-export default ErrorRecordContext;
