+++ src/user/provider.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import {
   createContext,
   Dispatch,
